# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "VSTR Management System API",
        "description": "Visitor Management System API Documentation",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Local Server"
        },
        {
            "url": "https://api.vstrmgmt.com",
            "description": "Production Server"
        },
        {
            "url": "https://staging-api.vstrmgmt.com",
            "description": "Staging Server"
        }
    ],
    "paths": {
...
        "/super-admin/admins": {
            "get": {
                "tags": [
                    "Admins"
                ],
                "summary": "List admin users",
                "description": "List administrators (super-admin)",
                "operationId": "listAdmins",
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Admins"
                ],
                "summary": "Create admin user",
                "description": "Create a new admin user",
                "operationId": "createAdmin",
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/super-admin/admins/{admin}": {
            "get": {
                "tags": [
                    "Admins"
                ],
                "summary": "Show admin user",
                "description": "Retrieve admin user details",
                "operationId": "showAdmin",
                "parameters": [
                    {
                        "name": "admin",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Admins"
                ],
                "summary": "Update admin user",
                "description": "Update an admin user",
                "operationId": "updateAdmin",
                "parameters": [
                    {
                        "name": "admin",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Admins"
                ],
                "summary": "Delete admin user",
                "description": "Delete an admin user",
                "operationId": "deleteAdmin",
                "parameters": [
                    {
                        "name": "admin",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/super-admin/admins/{admin}/locations": {
            "put": {
                "tags": [
                    "Admins"
                ],
                "summary": "Assign admin locations",
                "description": "Assign locations to an admin user",
                "operationId": "assignAdminLocations",
                "parameters": [
                    {
                        "name": "admin",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}